POC: Refactors the standalone init to conform to the webhook more
